+++ NEWS.md
@@ -1,3 +1,5 @@
+# tensorflow (development version)
+
 # tensorflow 2.7.0
 
 - Default Tensorflow version installed by `install_tensorflow()` updated to 2.7